Update sketch paths in "Compile Examples" CI workflow
Due to the library having examples specific to the MKR and Nano boards, the CI workflow must be configured to compile
only the examples of that board. The examples were moved after the time the workflow was written without updating the
workflow, resulting in spurious workflow run failures as opposed to the legitimate failures caused by this library's
various bugs.