Correct library reference URL in metadata
The library.properties `url` field provides a location for users to find information about the library. The automatically generated library reference pages are accessed via a URL based on the library `name` value converted to all lower case. Use of the real library name with upper case letters results in a 404.